code fixes based on comments
