+++ pkg/istio/rbac/common/util.go
@@ -24,32 +24,6 @@ func ParseRoleFQDN(domainName zms.DomainName, roleFQDN string) (string, error) {
 	return roleName, nil
 }
 
-func RoleToSpiffe(athenzDomainName string, roleName string) (string, error) {
-	if len(athenzDomainName) == 0 {
-		return "", fmt.Errorf("athenzDomainName is empty")
-	}
-	if len(roleName) == 0 {
-		return "", fmt.Errorf("roleName is empty")
-	}
-	spiffeName := fmt.Sprintf("%s/ra/%s", athenzDomainName, roleName)
-	return spiffeName, nil
-}
-
-// PrincipalToSpiffe converts the Athenz principal into a SPIFFE compliant format
-// e.g. client-domain.frontend.some-app -> client-domain.frontend/sa/some-app
-func PrincipalToSpiffe(principal string) (string, error) {
-	if len(principal) == 0 {
-		return "", fmt.Errorf("principal is empty")
-	}
-	i := strings.LastIndex(principal, ".")
-	if i < 0 {
-		return "", fmt.Errorf("principal:%s is not of the format <Athenz-domain>.<Athenz-service>", principal)
-	}
-	memberDomain, memberService := principal[:i], principal[i+1:]
-	spiffeName := fmt.Sprintf("%s/sa/%s", memberDomain, memberService)
-	return spiffeName, nil
-}
-
 // NewConfig returns a new model.Config resource for the passed-in type with the given namespace/name and spec
 func NewConfig(schema collection.Schema, namespace string, name string, spec proto.Message) model.Config {
 	meta := model.ConfigMeta{
